+++ Planarian.Web/src/Modules/Account/Services/AccountService.ts
@@ -17,10 +17,9 @@ import { EntranceDryRun } from "../../Import/Models/EntranceDryRun";
 import { FileImportResult } from "../../Import/Models/FileUploadresult";
 import { ArchiveListItemVm } from "../Models/Archive/ArchiveListItemVm";
 import { ArchiveProgressVm } from "../Models/Archive/ArchiveProgressVm";
-import { isNullOrWhiteSpace } from "../../../Shared/Helpers/StringHelpers";
 import { FileAccessUrlVm } from "../../Files/Services/FileService";
 import {
-  CreateImportFileUploadSessionRequest,
+  ImportFileRequest,
   ImportFileUploadSession,
 } from "../../Import/Models/ImportFileUploadSession";
 
@@ -146,45 +145,8 @@ const AccountService = {
     return response.data;
   },
 
-  async ImportFile(
-    file: string | Blob | RcFile,
-    uuid: string,
-    delmiterRegex: string,
-    idRegex: string,
-    ignoreDuplicates: boolean = false,
-    onProgress: (progressEvent: AxiosProgressEvent) => void,
-    signal?: AbortSignal
-  ): Promise<FileImportResult> {
-    const formData = new FormData();
-    formData.append("file", file);
-
-    const config: AxiosRequestConfig = {
-      headers: {
-        "Content-Type": "multipart/form-data",
-      },
-      onUploadProgress: onProgress, // Set the onUploadProgress callback
-      signal,
-    };
-
-    let regexQueryStringUrlSafe = `delimiterRegex=${encodeURIComponent(
-      delmiterRegex
-    )}&idRegex=${encodeURIComponent(
-      idRegex
-    )}&ignoreDuplicates=${ignoreDuplicates}`;
-
-    if (!isNullOrWhiteSpace(uuid)) {
-      regexQueryStringUrlSafe += `&uuid=${encodeURIComponent(uuid)}`;
-    }
-
-    const response = await HttpClient.post<FileImportResult>(
-      `${baseUrl}/import/file?${regexQueryStringUrlSafe}`,
-      formData,
-      config
-    );
-    return response.data;
-  },
   async CreateImportFileUploadSession(
-    request: CreateImportFileUploadSessionRequest
+    request: ImportFileRequest
   ): Promise<ImportFileUploadSession> {
     const response = await HttpClient.post<ImportFileUploadSession>(
       `${baseUrl}/import/file/session`,
