+++ Planarian/Planarian/Modules/Account/Archive/Models/ArchiveProgressVm.cs
@@ -1,6 +1,6 @@
 using Planarian.Shared.Models;
 
-namespace Planarian.Modules.Account.Model;
+namespace Planarian.Modules.Account.Archive.Models;
 
 public class ArchiveProgressVm : ProgressVm
 {
