+++ Planarian/Planarian/Modules/Account/Model/CreateAccountVm.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using Planarian.Model.Shared;
 
-namespace Planarian.Modules.Account.Controller;
+namespace Planarian.Modules.Account.Model;
 
 public class CreateAccountVm
 {
