+++ Planarian/Planarian/Modules/Account/Services/AccountService.cs
@@ -5,13 +5,13 @@ using Planarian.Library.Exceptions;
 using Planarian.Model.Database.Entities;
 using Planarian.Model.Database.Entities.RidgeWalker;
 using Planarian.Model.Shared;
+using Planarian.Modules.Account.Archive.Models;
 using Planarian.Modules.Account.Archive.Services;
 using Planarian.Modules.Account.Controller;
 using Planarian.Modules.Account.Model;
 using Planarian.Modules.Authentication.Services;
 using Planarian.Modules.Account.Repositories;
 using Planarian.Modules.Caves.Services;
-using Planarian.Modules.FeatureSettings.Repositories;
 using Planarian.Modules.Files.Models;
 using Planarian.Modules.Files.Repositories;
 using Planarian.Modules.Files.Services;
