+++ Planarian/Planarian/Modules/Caves/Services/CaveService.cs
@@ -15,9 +15,9 @@ using Planarian.Library.Options;
 using Planarian.Model.Database.Entities;
 using Planarian.Model.Database.Entities.RidgeWalker;
 using Planarian.Model.Shared;
+using Planarian.Modules.Account.Repositories;
 using Planarian.Modules.Caves.Models;
 using Planarian.Modules.Caves.Repositories;
-using Planarian.Modules.FeatureSettings.Repositories;
 using Planarian.Modules.Files.Repositories;
 using Planarian.Modules.Files.Services;
 using Planarian.Modules.Query.Extensions;
