+++ Planarian/Planarian/Modules/Map/Models/CoordinateDto.cs
@@ -1,4 +1,4 @@
-namespace Planarian.Modules.Map.Controllers;
+namespace Planarian.Modules.Map.Models;
 
 public class CoordinateDto
 {
